requirements.txt
streamlit==1.32.0
numpy==1.26.4
pandas==2.2.1
plotly==5.20.0
folium==0.16.0
streamlit-folium==0.20.0
tensorflow==2.15.0
scikit-learn==1.4.1
Pillow==10.2.0
matplotlib==3.8.3
